requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-datetime-widget==0.9.3
django-datetime-widget2==0.9.5
django-tempus-dominus==5.1.2.13
django-utils-six==2.0
docutils==0.16
idna==2.10
importlib-metadata==1.7.0
Kivy-Garden==0.1.4
Pillow==7.2.0
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
